finish init
start on internal gen (untested since no set and wont work without it)

+++ cmd/file_source.go
@@ -8,9 +8,11 @@ import (
 )
 
 func generateActivateSource(env string) error {
-	activateContent := `temp_src=$(%s zzz_INTERNAL_gen %s)
-source $temp_src
-rm $temp_src
+	activateContent := `temp_src=$(%s zzz_INTERNAL_gen -e %s)
+if [ $? -lt 1 ]; then
+    source $temp_src
+    rm $temp_src
+fi
 `
 
 	err := os.WriteFile(path.Join(".epicenv", env, "activate"), []byte(fmt.Sprintf(activateContent, lo.Ternary(os.Getenv("DEV") == "1", "go run .", "epicenv"), env)), 0777)
